Move DiscreteMarkovChain marginalization into its own module
Relocate the HMM marginalization logp, op subclass, and rewriter from
enumerable.py into a dedicated hmm.py, and the HMM marginal-logp tests
into test_hmm.py. The shared rewriter boilerplate is factored into
build_enumerable_marginal_rv, so the finite-discrete and Markov-chain
rewriters can live in their respective modules. Pure refactor.
