+++ pymc_extras/model/marginal/distributions/__init__.py
@@ -1,3 +1,4 @@
+import pymc_extras.model.marginal.distributions.discrete_markov_chain
 import pymc_extras.model.marginal.distributions.enumerable
 import pymc_extras.model.marginal.distributions.laplace
 import pymc_extras.model.marginal.distributions.normal  # noqa: F401
